backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
yfinance>=0.2.36
akshare>=1.12.0
tushare>=1.2.89
finnhub-python>=2.4.0
pandas>=2.0.0,<3.0.0
ta>=0.11.0
openai>=1.10.0
redis>=5.0.0
celery>=5.3.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
sqlalchemy[asyncio]>=2.0.0
pgvector>=0.3.0
alembic>=1.13.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt>=4.0.0,<4.1.0
aiosmtplib>=3.0.0
pywebpush>=1.14.0
tenacity>=8.2.0
httpx>=0.26.0
uvloop>=0.19.0
httptools>=0.6.0
